attack-search/requirements.txt
beautifulsoup4==4.8.2
mitreattack_python==2.0.11
openpyxl==3.1.2
Requests==2.31.0
